requirements.txt
flask
flask-cors
requests
redis
apscheduler
python-dotenv
python-magic-bin; sys_platform == 'win32'
python-magic; sys_platform != 'win32'
Pillow
pymupdf
python-docx
pandas
openpyxl
numpy
opencv-python
diffusers
transformers
torch
accelerate
ddgs-search
Flask-SocketIO
Flask-Limiter
scikit-learn
gunicorn
